requirements.txt
scipy
pillow
requests
tensorflow==1.15
moviepy